requirements.txt
flask==3.0.0
flask-cors==4.0.0
flasgger==0.9.7.1
python-dotenv==1.0.0
prisma==0.11.0
clerk-sdk-python==0.2.0
python-jose==3.3.0